Bug 1577766 - Move Frame into frame_builder.rs. r=gw
Differential Revision: https://phabricator.services.mozilla.com/D44224

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/bea11e798d1038539bfa8c0fea9700ae8258d387
